error handling/swallowing (#58)
#### Motivation:
- throw and catch custom errors, but does not surface them at the moment
- resolves #32 
- need another ticket to surface errors

#### Changelist:
- GameSizeOutOfRange exception
- QuestSizeIncorrect exception

#### Intended Effects:
- no more crashing

#### Proof of Testing:
- yes

+++ webapp/game/models.py
@@ -1,6 +1,8 @@
 from django.db import models
 from django.contrib.auth.models import AbstractUser
 
+from game.exceptions import QuestSizeIncorrect
+
 
 class AvalonUser(AbstractUser):
     pass
@@ -189,8 +191,9 @@ class AvalonQuest(models.Model):
         return AvalonQuestMember.objects.filter(quest=self)
 
     def reset_members(self, game_user_ids):
-        assert len(game_user_ids) == self.num_players, 'quest size is incorrect'
-        # KW: TODO persist previous tries
+        if len(game_user_ids) != self.num_players:
+            raise QuestSizeIncorrect
+
         AvalonQuestMember.objects.filter(quest=self).delete()
 
         for user_id in game_user_ids:
